+++ src/main/java/br/com/infocedro/promocontrol/core/exception/PeriodoEncerramentoPendenteInvalidoException.java
@@ -1,0 +1,8 @@
+package br.com.infocedro.promocontrol.core.exception;
+
+public class PeriodoEncerramentoPendenteInvalidoException extends BadRequestBusinessException {
+
+    public PeriodoEncerramentoPendenteInvalidoException() {
+        super("Periodo informado para entradas sem saida e invalido");
+    }
+}